+++ src/frontend/components/ProductCard/ProductCard.tsx
@@ -5,9 +5,17 @@ import { CypressFields } from '../../utils/Cypress';
 import { Product } from '../../protos/demo';
 import ProductPrice from '../ProductPrice';
 import * as S from './ProductCard.styled';
+import { useState, useEffect } from 'react';
+import { RequestInfo } from 'undici-types';
+import { useNumberFlagValue } from '@openfeature/react-sdk';
 
 interface IProps {
   product: Product;
+}
+
+async function getImageWithHeaders(url: RequestInfo, headers: Record<string, string>) {
+  const res = await fetch(url, { headers });
+  return await res.blob();
 }
 
 const ProductCard = ({
@@ -22,10 +30,20 @@ const ProductCard = ({
     },
   },
 }: IProps) => {
+  const imageSlowLoad = useNumberFlagValue('imageSlowLoad', 0);
+  const [imageSrc, setImageSrc] = useState<string>('');
+
+  useEffect(() => {
+    const headers = { 'x-envoy-fault-delay-request': imageSlowLoad.toString(), 'Cache-Control': 'no-cache' };
+    getImageWithHeaders('/images/products/' + picture, headers).then(blob => {
+      setImageSrc(URL.createObjectURL(blob));
+    });
+  }, [picture, imageSlowLoad]);
+
   return (
     <S.Link href={`/product/${id}`}>
       <S.ProductCard data-cy={CypressFields.ProductCard}>
-        <S.Image $src={"/images/products/" + picture} />
+        <S.Image $src={imageSrc} />
         <div>
           <S.ProductName>{name}</S.ProductName>
           <S.ProductPrice>